fix circular dependecy in initialFormValues, better credits, etc
