+++ src/common/providers/ResultProvider.ts
