+++ src/config/form-config/index.ts
@@ -1,31 +1,13 @@
 import { useTranslation } from "@/common/hooks/useLanguage"
-import type { InputLabel, PassCreationRules } from "@/models"
+import type { InputLabel } from "@/models"
 import { Language } from "@/models/translations"
 
 /** The version number for IDB stores to allow for sunsetting KV pairs */
 export const STORE_VERSION = 1 as const
+export const FORM_STATE_KEY = `formState-V${STORE_VERSION}`
+
 export const supportedLanguages = Object.values(Language)
 export const inputFieldMaxLength = 8
-export const defaultSliderValue = 3
-export const defaultFormValues: PassCreationRules = {
-  words: {
-    inputType: "radio",
-    selected: true,
-  },
-  uppercase: {
-    inputType: "checkbox",
-    selected: true,
-  },
-  numbers: {
-    inputType: "checkbox",
-    selected: true,
-  },
-  randomChars: {
-    inputType: "input",
-    value: "-",
-    selected: true,
-  },
-}
 
 export function labelForCheckbox(option: InputLabel) {
   const { t } = useTranslation()
