+++ src/styles/App.css
@@ -7,15 +7,25 @@
   height: 6px;
   border: 4px solid rgba(0, 0, 0, 0);
   background-clip: padding-box;
-  background-color: rgba(224, 224, 224, 0.2);
+  background-color: rgba(40, 40, 40, 0.7);
   -webkit-border-radius: 10px;
   -webkit-box-shadow: inset -1px -1px 0px rgba(0, 0, 0, 0.05),
     inset 1px 1px 0px rgba(0, 0, 0, 0.05);
   transition: color 900ms ease-in-out;
 }
 
+@media (prefers-color-scheme: dark) {
+  ::-webkit-scrollbar-thumb {
+    background-color: rgba(224, 224, 224, 0.2);
+  }
+
+  ::-webkit-scrollbar-thumb:hover {
+    background-color: rgba(224, 224, 224, 0.4);
+  }
+}
+
 ::-webkit-scrollbar-thumb:hover {
-  background-color: rgba(224, 224, 224, 0.4);
+  background-color: rgba(40, 40, 40, 0.9);
   -webkit-border-radius: 7px;
   -webkit-box-shadow: inset -1px -1px 0px rgba(0, 0, 0, 0.05),
     inset 1px 1px 0px rgba(0, 0, 0, 0.05);
@@ -33,21 +43,26 @@
 }
 
 .main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
   padding-top: 4rem;
   padding-bottom: 6rem;
   padding-left: 0.75rem;
   padding-right: 0.75rem;
+  min-height: 100dvh;
   width: 100vw;
+  margin: auto;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
 }
 
 .wrapper {
+  margin: 0 auto;
   display: flex;
   flex-direction: column;
+  justify-content: space-between;
   gap: 1.5rem;
   width: clamp(300px, 95%, var(--max-width));
+  height: 100%;
   max-width: var(--max-width);
 }
